+++ Formula/wadi.rb
@@ -2,7 +2,7 @@ class Wadi < Formula
   desc "Dune-free OCaml workspace toolbox"
   homepage "https://github.com/improvingjef/wadi"
   url "https://github.com/improvingjef/wadi/releases/download/v0.1.0/wadi-0.1.0-source.tar.gz"
-  sha256 "1bd002999335e2c1abb4ffe13b938b57c9b9b06bf7a3c1802ddb6b5c590501a2"
+  sha256 "0834b4efd8e0eb22c54d8ca635fad844e4c730604e59a561f331b54ba0513aaa"
   license "MIT"
 
   depends_on "ocaml"